golang: Use libtool
Unlike Rust, Go doesn't allow you to create libtool libraries
directly so we have to "get creative" - which is another way
of saying "build a pretty gross hack".

It seems that cgo is already emitting PIC object files, so we
should be good on that side, but libtool still needs a .la file
to operate; following what

  https://github.com/kanru/libtool-rs/blob/master/src/lib.rs

does, generate a simple .la file and set up the necessary
symbolic links. This seems to be enough to make libtool happy.