Clean up Makefile
In commit 762a3bd11e99 the file was moved to the upper directory, which makes
sense when you want to just run `make` in the cloned repo.  But after that some
commits added some more logic and it became just a script to be called with no
reason for it being a Makefile.

To make it a bit more readable and also not to run everything every time, we
first need to move it back.  That also removes the need for some `&& \` line
endings.  To keep the possibility of running `make` in the top level directory,
a redirecting Makefile is kept there.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>