Restructure capi/Makefile
To make it a bit more readable and also not to run everything every single time
make is executed, create some relevant targets.

This still keeps running go every time, as requested in commit 762a3bd11e99, and
because we cannot check whether the output changed, everything related to the go
build is run every time as well.  But it makes it at least a bit nicer.

This does not go into extremes like having a target for the `.libs` directory
etc., it is merely making it nicer to read, at least for some people.  Take it
as a mere suggestion.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>