c/rust: Introduce error handling
It works in a way that's intentionally very similar to what GLib
does with GError, though there are some differences (at least for
the moment).

A dummy error type is included for demonstration purposes, but
we'll get rid of it as soon as the first real error type is
exposed via the bindings.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>