c/go: Split source files
Having all the glue code and all the wrappers in a single file
each is unmaintainable in the long, and even not so long, term.

Unfortunately it doesn't look like we can use directories to keep
our source files neatly organized (cgo doesn't like that), but at
least we can have separate files even if they all end up in the
same directory.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>