go/rust: Tweak LDFLAGS
When linking regular C programs against the C -> Rust bindings we
include '-pthread -lm', so it makes sense to include them when
building the Go -> Rust bindings as well - despite the fact that
everything seems to work just fine as-is.

Since we're at it, we can also use a shorter library search path
which, as a bonus, doesn't hardcode the fact that we're only doing
debug builds at the moment - something that's bound to eventually
change.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>