c/go: Use assert() in the wrapper code
In some scenarios it's fine to pass a NULL pointer to the glue code,
but when dealing with objects specifically we can't permit that as we
need to unwrap the C object to obtain a Go object reference. Ensure
users do not pass NULL pointers by assert()ing.

Note that the free functions still accept NULL for convenience.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>