rust/native: Remove embedded copy of closefds crate
The copy contains a few changes and fixes on top of what's on
crates.io, so we can't quite use the vanilla version; that said,
simply copying the code over is a bit yucky.

A slightly cleaner solution is to have a clone of the upstream
repository under our organization and make the dependency point
to a specific branch containing the additional patches.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>