c/rust: Properly perform io::Error -> errno conversion
When running the following code,

  let err1 = io::Error::from_raw_os_error(22);

  assert!(err1.kind() == io::ErrorKind::InvalidInput);
  assert!(err1.raw_os_error() == Some(22));

  let err2 = io::Error::new(io::ErrorKind::InvalidInput, "");

  assert!(err2.kind() == io::ErrorKind::InvalidInput);
  assert!(err2.raw_os_error() == Some(22));

the last assert!() will fail: this is because Rust will refuse to
perform a lossy conversion between an error that was not originally
created from an OS error and an errno, even in cases such as this
one where the conversion is well-defined.

It makes sense that Rust wouldn't allow the lossy conversion in the
general case, but as far as we're concerned it's actually perfectly
safe, so implement the mapping ourselves: this will allow C users
to just compare the return value of virtblocks_error_get_code() with
values defined in errno.h.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>