c/go: Introduce types.Array
This is a generic container for C-accessible types, and will be
used when we need to return eg. an array of strings to C.

It complies with cgo's memory sharing constraints, and is also
somewhat more convenient to use than a regular C array because it
removes the need to track the length with a separate variable and
supports VIRTBLOCKS_AUTOPTR().

The drawback is that C code using it ends up being more verbose,
but such is life.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>