Flatten repository structure
Now that we only have a single native implementation and a single
set of language bindings, we can make the repository structure
flatter.

We could arguably move the native Go implementation to the top
level directory, but that'd make things a tiny bit awkward with
the Makefile, so for now it's staying in its own directory. We
might very well reconsider this decision later.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>