examples: Adjust command examples
We want the Go and C examples to be as close as possible. The C
bindings don't yet expose all features, but we'll take care of
that eventually.

The amount of sleep is also reduced so that running the test suite
takes less time, and the test script is changed so that it writes
both on stdout and stderr.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>