c: Rework error handling
Instead of storing the original error in a VirtBlocksError and
then extracting the relevant part when virtblocks_error_get_*() is
called, perform the conversion upfront and store C types directly
in the structure.

Doing this has a couple of drawbacks: it breaks the rule we've
followed so far of keeping everything under pkg/ pure Go, and more
importantly results in allocating upfront C strings that we might
never get to use.

The advantage is that, now that we no longer need to worry about
freeing the value returned by virtblocks_error_get_message(), with
a couple of very simple additional macros we can remove a massive
amount of boilerplate from C code that needs to deal with errors,
bringing it almost to the same level as the corresponding Go code.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>