c: Further simplify error handling
When exposing any other type to C we're pretty strict: pass a NULL
pointer, get a panic.

Errors, however, are so pervasive to the API that requiring them to
be handled this strictly results in more verbose code or having to
work around the unpleasantness with macros.

Let's take a more pragmatic approach: any API can be called on a
NULL error, and the result will be bogus data - but no panics.

This is fine because the usage pattern is supposed to be along the
lines of

  val = get(&err);
  if (err) {
      printf("err: %s\n", virtblocks_error_get_message(err));
      return 1;
  }
  use(val);

so in most situations you're not going to call APIs on an error
until you've already confirmed it's non-NULL, and in the few
remaining cases (mostly debugging) it's very convenient to be able
to use error APIs unconditionally.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>