go/pkg/vm: Implement disk slots
Each VM can now have multiple disks, each one attached to a
specific disk slot. For now we're not actually using the new
information as intendend, but merely validating it to make sure
it makes sense.

In the examples and demo launcher we use somewhat arbitrary slot
numbers to ensure things don't fall completely when we have more
than a single slot.

The C bindings are also affected by this change.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>